Refactor: ApplyListService -> ApplyService로 변경
