--- src/main/java/com/mycompany/metamong/service/ApplyListService.java
@@ -1,34 +1,0 @@
-package com.mycompany.metamong.service;
-
-import java.util.List;
-import java.util.Map;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.mycompany.metamong.daoMain.ApplyListDao;
-import com.mycompany.metamong.dto.applyList.ApplyListDto;
-import com.mycompany.metamong.dto.applyList.ApplyTableDeatilDto;
-import com.mycompany.metamong.dto.table.ApplyTableDto;
-
-@Service
-public class ApplyListService {
-	@Autowired
-	private ApplyListDao applyListDao;
-
-	public void addApplyList(ApplyListDto apply) {
-		applyListDao.insertApplyList(apply);	
-	}
-
-	public List<ApplyTableDto> getApplyTableList() {
-		return applyListDao.selectApplyTableList();
-	}
-
-	public List<ApplyTableDto> getApplyTableSearch(Map<String, String> form) {
-		return applyListDao.selectApplyTableSearch(form);
-	}
-
-	public ApplyTableDeatilDto getTableListDetail(int applyNo) {
-		return applyListDao.selectTableListDetail(applyNo);
-	}
-}
