+++ src/main/java/com/mycompany/metamong/service/ApplyService.java
@@ -1,24 +1,47 @@
 package com.mycompany.metamong.service;
 
+import java.util.List;
+import java.util.Map;
+
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
+import com.mycompany.metamong.daoMain.ApplyListDao;
+import com.mycompany.metamong.daoMain.IndexDao;
 import com.mycompany.metamong.dto.applyList.ApplyListDto;
+import com.mycompany.metamong.dto.applyList.ApplyTableDeatilDto;
 import com.mycompany.metamong.dto.index.ApplyIndexDto;
+import com.mycompany.metamong.dto.table.ApplyTableDto;
 
 @Service
 public class ApplyService {
 	@Autowired
-	private ApplyListService applyListService;
+	private ApplyListDao applyListDao;
 	@Autowired
-	private IndexService indexService;
+	private IndexDao indexDao;
+
+	public void addApplyList(ApplyListDto apply) {
+		applyListDao.insertApplyList(apply);	
+	}
+
+	public List<ApplyTableDto> getApplyTableList() {
+		return applyListDao.selectApplyTableList();
+	}
+
+	public List<ApplyTableDto> getApplyTableSearch(Map<String, String> form) {
+		return applyListDao.selectApplyTableSearch(form);
+	}
+
+	public ApplyTableDeatilDto getTableListDetail(int applyNo) {
+		return applyListDao.selectTableListDetail(applyNo);
+	}
 	
 	@Transactional
 	public void addApplyIndex(ApplyListDto applyListDto, ApplyIndexDto applyIndexDto) {
-		applyListService.addApplyList(applyListDto);
+		applyListDao.insertApplyList(applyListDto);
 		applyIndexDto.setApplyNo(applyListDto.getApplyNo());
-		indexService.addApplyIndex(applyIndexDto);
+		indexDao.insertApplyIndex(applyIndexDto);
 	}
-	
 }
